Made things tidier (moved to files + renames) + added doc
